+++ source/lib/data/EventsHeader.h
@@ -93,7 +93,7 @@ namespace blitzortung {
 	Event::AP createEvent(std::iostream& stream, const Format& dataFormat) const;
 
 	//! return size of header in bytes
-	unsigned int getSize() const;
+	size_t getSize() const;
     };
 
     //! stream output operator
